+++ Assets/Scripts/Mechanics/AirPortGranter.cs
@@ -53,7 +53,7 @@ public class AirPortGranter : SavableMonoBehaviour
         Used = true;
     }
 
-    static short key_used = 0;
+    static byte key_used = 0;
     public override SavableObject CurrentState
     {
         get => new SavableObject(this,
